# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BFC Fast Delivery API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: restaurants
    description: Restaurant operations
  - name: menu
    description: Menu operations
  - name: orders
    description: Order operations
  - name: categories
    description: Category operations
  - name: admin
    description: Admin operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Register a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupBody"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login to account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                required:
                  - message

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all food categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

  /restaurants:
    get:
      operationId: listRestaurants
      tags: [restaurants]
      summary: List all restaurants
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of restaurants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Restaurant"

  /restaurants/featured:
    get:
      operationId: listFeaturedRestaurants
      tags: [restaurants]
      summary: List featured restaurants for homepage
      responses:
        "200":
          description: Featured restaurants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Restaurant"

  /restaurants/{restaurantId}:
    get:
      operationId: getRestaurant
      tags: [restaurants]
      summary: Get a restaurant by ID
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Restaurant details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restaurant"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /restaurants/{restaurantId}/menu:
    get:
      operationId: getRestaurantMenu
      tags: [menu]
      summary: Get menu items for a restaurant
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Menu items grouped by category
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuSection"

  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders for current user
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /orders/{orderId}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /stats/overview:
    get:
      operationId: getStatsOverview
      tags: [restaurants]
      summary: Get platform stats overview
      responses:
        "200":
          description: Platform statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsOverview"

  /admin/orders:
    get:
      operationId: adminListOrders
      tags: [admin]
      summary: Admin - list all orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: All orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

  /admin/orders/{orderId}/status:
    patch:
      operationId: adminUpdateOrderStatus
      tags: [admin]
      summary: Admin - update order status
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusBody"
      responses:
        "200":
          description: Order updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/restaurants:
    post:
      operationId: adminCreateRestaurant
      tags: [admin]
      summary: Admin - create a restaurant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRestaurantBody"
      responses:
        "201":
          description: Restaurant created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restaurant"

  /admin/restaurants/{restaurantId}:
    patch:
      operationId: adminUpdateRestaurant
      tags: [admin]
      summary: Admin - update a restaurant
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRestaurantBody"
      responses:
        "200":
          description: Restaurant updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restaurant"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: adminDeleteRestaurant
      tags: [admin]
      summary: Admin - delete a restaurant
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /admin/menu-items:
    post:
      operationId: adminCreateMenuItem
      tags: [admin]
      summary: Admin - create a menu item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMenuItemBody"
      responses:
        "201":
          description: Menu item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"

  /admin/menu-items/{menuItemId}:
    patch:
      operationId: adminUpdateMenuItem
      tags: [admin]
      summary: Admin - update a menu item
      parameters:
        - name: menuItemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMenuItemBody"
      responses:
        "200":
          description: Menu item updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: adminDeleteMenuItem
      tags: [admin]
      summary: Admin - delete a menu item
      parameters:
        - name: menuItemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /admin/stats:
    get:
      operationId: adminGetStats
      tags: [admin]
      summary: Admin - get detailed platform stats
      responses:
        "200":
          description: Admin stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    AuthUser:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [customer, admin]
      required:
        - id
        - name
        - email
        - role

    SignupBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [customer, admin]
      required:
        - name
        - email
        - password

    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password

    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        icon:
          type: string
        slug:
          type: string
      required:
        - id
        - name
        - icon
        - slug

    Restaurant:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        rating:
          type: number
        reviewCount:
          type: integer
        deliveryTime:
          type: string
        minimumOrder:
          type: number
        categoryId:
          type: integer
        categoryName:
          type: string
        address:
          type: string
        isOpen:
          type: boolean
        isFeatured:
          type: boolean
      required:
        - id
        - name
        - description
        - imageUrl
        - rating
        - reviewCount
        - deliveryTime
        - minimumOrder
        - categoryId
        - categoryName
        - address
        - isOpen
        - isFeatured

    MenuItem:
      type: object
      properties:
        id:
          type: integer
        restaurantId:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        imageUrl:
          type: string
        category:
          type: string
        isAvailable:
          type: boolean
        isPopular:
          type: boolean
      required:
        - id
        - restaurantId
        - name
        - description
        - price
        - imageUrl
        - category
        - isAvailable
        - isPopular

    MenuSection:
      type: object
      properties:
        category:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/MenuItem"
      required:
        - category
        - items

    OrderItem:
      type: object
      properties:
        menuItemId:
          type: integer
        name:
          type: string
        price:
          type: number
        quantity:
          type: integer
      required:
        - menuItemId
        - name
        - price
        - quantity

    Order:
      type: object
      properties:
        id:
          type: integer
        restaurantId:
          type: integer
        restaurantName:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        totalAmount:
          type: number
        deliveryAddress:
          type: string
        customerName:
          type: string
        customerPhone:
          type: string
        status:
          type: string
          enum: [pending, confirmed, preparing, out_for_delivery, delivered, cancelled]
        estimatedDelivery:
          type: string
        createdAt:
          type: string
        userId:
          type: integer
          nullable: true
      required:
        - id
        - restaurantId
        - restaurantName
        - items
        - totalAmount
        - deliveryAddress
        - customerName
        - customerPhone
        - status
        - estimatedDelivery
        - createdAt

    CreateOrderBody:
      type: object
      properties:
        restaurantId:
          type: integer
        items:
          type: array
          items:
            type: object
            properties:
              menuItemId:
                type: integer
              quantity:
                type: integer
            required:
              - menuItemId
              - quantity
        deliveryAddress:
          type: string
        customerName:
          type: string
        customerPhone:
          type: string
      required:
        - restaurantId
        - items
        - deliveryAddress
        - customerName
        - customerPhone

    UpdateOrderStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, preparing, out_for_delivery, delivered, cancelled]
      required:
        - status

    CreateRestaurantBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        rating:
          type: number
        reviewCount:
          type: integer
        deliveryTime:
          type: string
        minimumOrder:
          type: number
        categoryId:
          type: integer
        categoryName:
          type: string
        address:
          type: string
        isOpen:
          type: boolean
        isFeatured:
          type: boolean
      required:
        - name
        - description
        - imageUrl
        - deliveryTime
        - minimumOrder
        - categoryId
        - categoryName
        - address

    UpdateRestaurantBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        rating:
          type: number
        reviewCount:
          type: integer
        deliveryTime:
          type: string
        minimumOrder:
          type: number
        categoryId:
          type: integer
        categoryName:
          type: string
        address:
          type: string
        isOpen:
          type: boolean
        isFeatured:
          type: boolean

    CreateMenuItemBody:
      type: object
      properties:
        restaurantId:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        imageUrl:
          type: string
        category:
          type: string
        isAvailable:
          type: boolean
        isPopular:
          type: boolean
      required:
        - restaurantId
        - name
        - description
        - price
        - imageUrl
        - category

    UpdateMenuItemBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        imageUrl:
          type: string
        category:
          type: string
        isAvailable:
          type: boolean
        isPopular:
          type: boolean

    StatsOverview:
      type: object
      properties:
        totalRestaurants:
          type: integer
        totalOrders:
          type: integer
        totalCategories:
          type: integer
        avgDeliveryTime:
          type: string
      required:
        - totalRestaurants
        - totalOrders
        - totalCategories
        - avgDeliveryTime

    AdminStats:
      type: object
      properties:
        totalRestaurants:
          type: integer
        totalOrders:
          type: integer
        totalUsers:
          type: integer
        totalRevenue:
          type: number
        pendingOrders:
          type: integer
        deliveredOrders:
          type: integer
        todayOrders:
          type: integer
        totalMenuItems:
          type: integer
      required:
        - totalRestaurants
        - totalOrders
        - totalUsers
        - totalRevenue
        - pendingOrders
        - deliveredOrders
        - todayOrders
        - totalMenuItems