# lib/api-zod/src/generated/types/authUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BFC Fast Delivery API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AuthUserRole } from "./authUserRole";

export interface AuthUser {
  id: number;
  name: string;
  email: string;
  role: AuthUserRole;
}