# lib/api-zod/src/generated/types/createRestaurantBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BFC Fast Delivery API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateRestaurantBody {
  name: string;
  description: string;
  imageUrl: string;
  rating?: number;
  reviewCount?: number;
  deliveryTime: string;
  minimumOrder: number;
  categoryId: number;
  categoryName: string;
  address: string;
  isOpen?: boolean;
  isFeatured?: boolean;
}