# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BFC Fast Delivery API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from "./orderItem";
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  restaurantId: number;
  restaurantName: string;
  items: OrderItem[];
  totalAmount: number;
  deliveryAddress: string;
  customerName: string;
  customerPhone: string;
  status: OrderStatus;
  estimatedDelivery: string;
  createdAt: string;
  userId?: number | null;
}