# lib/api-zod/src/generated/types/signupBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BFC Fast Delivery API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SignupBodyRole } from "./signupBodyRole";

export interface SignupBody {
  name: string;
  email: string;
  password: string;
  role?: SignupBodyRole;
}