# lib/api-zod/src/generated/types/updateMenuItemBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BFC Fast Delivery API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateMenuItemBody {
  name?: string;
  description?: string;
  price?: number;
  imageUrl?: string;
  category?: string;
  isAvailable?: boolean;
  isPopular?: boolean;
}